+++ cedar-drt/src/logger.rs
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
+use log::warn;
 use std::time::{Duration, Instant};
 
 pub const RUST_SERIALIZATION_MSG: &str = "rust_serialization (ns) : ";
@@ -25,6 +25,21 @@ pub const JAVA_DESERIALIZATION_MSG: &str = "java_deserialization (ns) : ";
 pub const JAVA_AUTH_MSG: &str = "java_auth (ns) : ";
 pub const JAVA_VALIDATION_MSG: &str = "java_validation (ns) : ";
 pub const TOTAL_MSG: &str = "total (ns) : ";
+
+pub fn initialize_log() {
+    match env_logger::try_init() {
+        Ok(()) => (),
+        Err(e) => {
+            let msg = e.to_string();
+            if &msg == "attempted to set a logger after the logging system was already initialized"
+            {
+                // don't log that error, it's expected
+            } else {
+                warn!("SetLogError : {msg}");
+            }
+        }
+    };
+}
 
 pub fn time_function<X, F>(f: F) -> (X, Duration)
 where
